+++ src/main/kotlin/eu/europa/ec/eudi/sdjwt/vc/TypeMetadata.kt
@@ -213,116 +213,6 @@ value class SvgId(val value: String) {
     override fun toString() = value
 }
 
-@Serializable(with = ClaimPathSerializer::class)
-@JvmInline
-value class ClaimPath(val value: List<ClaimPathElement>) {
-
-    init {
-        require(value.isNotEmpty())
-    }
-
-    override fun toString(): String = value.toString()
-
-    operator fun plus(other: ClaimPathElement): ClaimPath =
-        ClaimPath(this.value + other)
-
-    operator fun plus(other: ClaimPath): ClaimPath =
-        ClaimPath(this.value + other.value)
-
-    /**
-     * Appends a wild-card indicator [ClaimPathElement.All]
-     */
-    fun all(): ClaimPath = this + ClaimPathElement.All
-
-    /**
-     * Appends an indexed path [ClaimPathElement.Indexed]
-     */
-    fun at(i: Int): ClaimPath = this + ClaimPathElement.Indexed(i)
-    fun attribute(name: String): ClaimPath = this + ClaimPathElement.Named(name)
-
-    fun head(): ClaimPathElement = value.first()
-    fun tail(): ClaimPath? {
-        val tailElements = value.drop(1)
-        return if (tailElements.isEmpty()) return null
-        else ClaimPath(tailElements)
-    }
-
-    operator fun component1(): ClaimPathElement = head()
-    operator fun component2(): ClaimPath? = tail()
-
-    companion object {
-        fun attribute(name: String): ClaimPath = ClaimPath(listOf(ClaimPathElement.Named(name)))
-    }
-}
-
-sealed interface ClaimPathElement {
-    data object All : ClaimPathElement {
-        override fun toString() = "null"
-    }
-
-    @JvmInline
-    value class Indexed(val value: Int) : ClaimPathElement {
-        init {
-            require(value >= 0) { "Index should be non-negative" }
-        }
-
-        override fun toString() = value.toString()
-    }
-
-    @JvmInline
-    value class Named(val value: String) : ClaimPathElement {
-        init {
-            require(value.isNotBlank()) { "Attribute must not be blank" }
-        }
-
-        override fun toString() = value
-    }
-}
-
-/**
- * Serializer for [ClaimPath]
- */
-object ClaimPathSerializer : KSerializer<ClaimPath> {
-
-    private fun claimPathElement(it: JsonPrimitive): ClaimPathElement =
-        when {
-            it is JsonNull -> ClaimPathElement.All
-            it.isString -> ClaimPathElement.Named(it.content)
-            it.intOrNull != null -> ClaimPathElement.Indexed(it.int)
-            else -> throw IllegalArgumentException("Only string, null, int can be used")
-        }
-
-    private fun claimPath(array: JsonArray): ClaimPath {
-        val elements = array.map {
-            require(it is JsonPrimitive)
-            claimPathElement(it)
-        }
-        return ClaimPath(elements)
-    }
-
-    private fun ClaimPath.toJson(): JsonArray = JsonArray(value.map { it.toJson() })
-
-    private fun ClaimPathElement.toJson(): JsonPrimitive = when (this) {
-        is ClaimPathElement.Named -> JsonPrimitive(value)
-        is ClaimPathElement.Indexed -> JsonPrimitive(value)
-        ClaimPathElement.All -> JsonNull
-    }
-
-    val arraySerializer = serializer<JsonArray>()
-
-    override val descriptor: SerialDescriptor = arraySerializer.descriptor
-
-    override fun serialize(encoder: Encoder, value: ClaimPath) {
-        val array = value.toJson()
-        arraySerializer.serialize(encoder, array)
-    }
-
-    override fun deserialize(decoder: Decoder): ClaimPath {
-        val array = arraySerializer.deserialize(decoder)
-        return claimPath(array)
-    }
-}
-
 @Serializable
 @JvmInline
 value class Display(val value: List<DisplayMetadata>) {
