+++ src/test/kotlin/eu/europa/ec/eudi/sdjwt/PidDevVerificationTest.kt
@@ -39,7 +39,7 @@ class PidDevVerificationTest : Printer {
             eyJhbGciOiAiRVMyNTYiLCAidHlwIjogInZjK3NkLWp3dCJ9.eyJfc2QiOiBbIjZfRUYxcV9ERURsMTVyOTRCQm1TaTZNalFaLWpJMVhLbzNfdW1xQVhlb0UiLCAiRnJ3WDN3R2xlbWVhdUVQMnRsSDBtS1lZYU9LOTVISVFqYmI0WUU3aWVIQSIsICJLNmJNTnpGMU5PeHk1STN4TkdPcGozSEVYVEtac3FXM0NtQjZLOWVBN1NzIiwgImdmN0lyMUJ5SnRRQ0tERm85dm45eGJ4UmowX2ZaTnFXZTB6eDJsNHRVcUUiLCAicDJQRExuRDB0ZnlVVy1acThPeGJjeG50RVUtQTVtODJWTGNQUm9TY0pyVSIsICJxSkliZjYxRzVnTEF6bzlfa1NlcFB3bTh0MEpHbngyOXlqRDVmejhPUnJFIl0sICJpc3MiOiAiaHR0cHM6Ly9kZXYuaXNzdWVyLmV1ZGl3LmRldiIsICJqdGkiOiAiMGEzOTk5NTAtYzliOS00OTFhLWI0OGItYjY2NDUxMTM1ZTBkIiwgImlhdCI6IDIwMDQ3LCAiZXhwIjogMjAxMzcsICJzdGF0dXMiOiAidmFsaWRhdGlvbiBzdGF0dXMgVVJMIiwgInZjdCI6ICJ1cm46ZXUuZXVyb3BhLmVjLmV1ZGkucGlkLjEiLCAiX3NkX2FsZyI6ICJzaGEtMjU2IiwgImNuZiI6IHsiandrIjogeyJrdHkiOiAiRUMiLCAiY3J2IjogIlAtMjU2IiwgIngiOiAid1V1UDJPbHdIZWZlRS1ZMTZXajdQSEF6WjBKQVF5ZXZxV01mZDUtS21LWSIsICJ5IjogIllXLWI4TzNVazNOVXJrOW9acEFUMWxhUGVBZ2lOUXdEY290V2l3QkZRNkUifX19.6Ub00iVL7rSpCN82wRUkT4kv-OgvTzOdIrUHeFKX29IbXKX4GpeJWd5hUmYCU8tOnWvhFcffONiGQgsWJIwYuw~WyJGZl9sak93Z25WTVFYazctbGk4N3F3IiwgImZhbWlseV9uYW1lIiwgIkZhbWlseSBOYW1lIFRlc3RlciJd~WyJ0Wk9aaEtOYnUyWm9DWHlua2dfQmR3IiwgImdpdmVuX25hbWUiLCAiR2l2ZW4gTmFtZSBJc3N1ZXIiXQ~WyJISkJqd2JvbmdnVnJ1QkJFT0xkenpBIiwgImJpcnRoZGF0ZSIsICIxOTk5LTExLTIwIl0~WyI0emdlSUNQQngzLUc3RDNyc3RkM2N3IiwgImlzc3VpbmdfYXV0aG9yaXR5IiwgIlRlc3QgUElEIGlzc3VlciJd~WyIyS0VzSEhBMjFoVjg5TkllcVhwbUpRIiwgImlzc3VpbmdfY291bnRyeSIsICJGQyJd~WyJ6Vm9rQlBmS052SmpwRU5kbG1rdUFBIiwgImFnZV9lcXVhbF9vcl9vdmVyIiwgeyIxOCI6IHRydWV9XQ~
     """.trimIndent()
 
-    @Test @Ignore
+    @Test
     fun testKotlin() = doTest(pid1, enableLogging = false)
 
     @Test @Ignore
