+++ src/test/kotlin/eu/europa/ec/eudi/sdjwt/vc/ClaimPathTest.kt
@@ -15,86 +15,11 @@
  */
 package eu.europa.ec.eudi.sdjwt.vc
 
-import kotlinx.serialization.json.*
+import kotlinx.serialization.json.Json
 import org.junit.jupiter.api.Test
-import kotlin.test.assertEquals
 
 class ClaimPathTest {
     val jsonSupport = Json { ignoreUnknownKeys = true }
-    val sampleJson = """
-        {
-          "vct": "https://betelgeuse.example.com/education_credential",
-          "name": "Arthur Dent",
-          "address": {
-            "street_address": "42 Market Street",
-            "city": "Milliways",
-            "postal_code": "12345"
-          },
-          "degrees": [
-            {
-              "type": "Bachelor of Science",
-              "university": "University of Betelgeuse"
-            },
-            {
-              "type": "Master of Science",
-              "university": "University of Betelgeuse"
-            }
-          ],
-          "nationalities": ["British", "Betelgeusian"]
-        }
-    """.trimIndent().let {
-        Json.parseToJsonElement(it).jsonObject
-    }
-
-    @Test
-    fun matchTopLevel() {
-        sampleJson.forEach { (attributeName, attribute) ->
-            assertMatchEquals(expected = attribute, path = ClaimPath.attribute(attributeName))
-        }
-    }
-    private fun assertMatchEquals(expected: JsonElement, path: ClaimPath) {
-        val actual = MatchClaimPath.Default.match(sampleJson, path).getOrThrow()
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun matchNestedAttribute() = assertMatchEquals(
-        expected = checkNotNull(checkNotNull(sampleJson["address"]).jsonObject["street_address"]),
-        path = ClaimPath.attribute("address").attribute("street_address"),
-    )
-
-    @Test
-    fun matchArray() = assertMatchEquals(
-        expected = checkNotNull(checkNotNull(sampleJson["degrees"]).jsonArray),
-        path = ClaimPath.attribute("degrees"),
-    )
-
-    @Test
-    fun matchAll() = assertMatchEquals(
-        expected = checkNotNull(checkNotNull(sampleJson["degrees"]).jsonArray),
-        path = ClaimPath.attribute("degrees").all(),
-    )
-
-    @Test
-    fun matchArrayElement() {
-        val degreesPath = ClaimPath.attribute("degrees")
-        val degrees = checkNotNull(sampleJson["degrees"]).jsonArray
-        degrees.forEachIndexed { index, degree ->
-            assertMatchEquals(expected = degree, path = degreesPath.at(index))
-        }
-    }
-
-    @Test
-    fun matchNestedOfArrayElement() = assertMatchEquals(
-        expected = run {
-            val degrees = checkNotNull(sampleJson["degrees"]).jsonArray
-            degrees.map { element ->
-                check(element is JsonObject)
-                checkNotNull(element["type"])
-            }.let(::JsonArray)
-        },
-        path = ClaimPath.attribute("degrees").all().attribute("type"),
-    )
 
     @Test
     fun `parsing happy path`() {
